==> 05_Evacuation_Tweet/code/00_Pull_California_Street_Names.ipynb <==
# Pull California Street Names
This notebook uses BeautifulSoup to scrape street names from [Geographic.org](https://geographic.org/streetview/usa/ca/). This data could be used for training a spaCy NLP function, but we did not get to it during our process.

from bs4 import BeautifulSoup
import requests
import pandas as pd
import time

url = 'https://geographic.org/streetview/usa/ca/'
tag = 'index2.html'

res = requests.get(url + tag)

soup = BeautifulSoup(res.content)

## Webscrape
Loop through each county and city to scrape the county, city, street name and zip code for each street.

t_0 = time.time()

counties_ul = soup.find('ul')
counties = counties_ul.find_all('li')

for i, county in enumerate(counties[30:]):
    county_anchor = county.find('a')
    county_tag = county_anchor.attrs['alt'].lower().replace(' ','_')
    county_name = county_anchor.text
    county_url = url + county_tag + '/index.html'
    county_res = requests.get(county_url)
    time.sleep(2)
    county_soup = BeautifulSoup(county_res.content)
    cities = county_soup.find('ul')
    for city in cities.find_all('li'):
        city_anchor = city.find('a')
        city_tag = city_anchor.attrs['href']
        city_name = city_anchor.attrs['alt']
        city_url = county_url.replace('index',city_name.lower())
        city_res = requests.get(city_url)
        time.sleep(2)
        city_soup = BeautifulSoup(city_res.content)
        streets = city_soup.find('ul')
        for street in streets.find_all('li'):
            road = {}
            road['county'] = county_name
            road['city'] = city_name
            road['street'] = street.find('a').attrs['alt']
            road['zip_code'] = street.text[-5:]
            roads.append(road)
    print(f"[{'#' * i}{' ' * (len(counties) - i)}] {county_name} {i/len(counties)}")
print(time.time() - t_0)
pd.DataFrame(roads)

pd.DataFrame(roads).to_csv('../data/california_street_names.csv', index=False)

# Summary
We scrapped all 268,000 California street names from [Geographic.org](https://geographic.org/streetview/usa/ca/).
