==> 05_Evacuation_Tweet/code/00_Wiki_Scrapes_Caltrans_Districts.ipynb <==
# CalTrans District Web Scrape

### OBJECTIVE

- This notebook will scrape districts and counties from the wikipedia url listed below.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import requests
from bs4 import BeautifulSoup

import selenium

url = 'https://en.wikipedia.org/wiki/California_Department_of_Transportation'

res = requests.get(url)

soup = BeautifulSoup(res.content)

tables = soup.find_all('table')

main_header = [th.text.lower().strip().replace(' ','_') for th in tables[1].find_all('th')]
main_header.append('table')
main_header

district_fires = []
for i, table in enumerate(tables):
    table_df = {}
    headers = {th.text.lower().strip().replace(' ','_'):index for index,th in enumerate(tables[1].find_all('th'))}

    for row in table.find_all('tr')[1:]:
        district_f = {}

        for j,feature in enumerate(main_header):
            try:
                district_f[feature] = row.find_all('td')[headers[feature]].text.strip()
            except:
                district_f[feature] = np.nan
                
        district_f['table'] = i
        district_fires.append(district_f)
pd.DataFrame(district_fires).to_csv('./Data/districts.csv',index=False)

df = pd.read_csv('./Data/districts.csv')

df.head()

df.shape

districts = df.loc[df['table'] == 1]

districts.head(50)

# Strip references from the cell
references = index.findAll("sup", {"class": "reference"})
if references:
    for ref in references:
        ref.extract()

# Strip sortkeys from the cell
sortkeys = index.findAll("span", {"class": "sortkey"})
if sortkeys:
    for ref in sortkeys:
        ref.extract()

# Strip footnotes from text and join into a single string
text_items = index.findAll(text=True)
no_footnotes = [text for text in text_items if text[0] != '[']
puretext = ''.join(no_footnotes)

# Replace non-breaking spaces with regular spaces and add quotes
puretext = puretext.replace('\xa0', ' ')

for col in districts.columns:
    print(col)
    districts[col] = districts[col].map(lambda x: re.sub(pattern='(\[\d+\])+$',repl='',string=x))

districts.head(50)

districts.to_csv('./Data/districts.csv',index=False)

### SUMMARY

- This notebook scrapes California district information from a wikipedia page. The columns include district name, counties, and headquarters.
