==> 05_Evacuation_Tweet/code/00_Wiki_Scrapes_Notable_Fires.ipynb <==
# Notable Fires Web Scrape

### OBJECTIVE

- This notebook will scrape California fire information from the wikipedia url listed below.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import requests
from bs4 import BeautifulSoup

import selenium
import regex as re

url = 'https://en.wikipedia.org/wiki/List_of_California_wildfires'

res = requests.get(url)

soup = BeautifulSoup(res.content)

tables = soup.find_all('table', class_= 'wikitable')

main_header = [th.text.lower().strip().replace(' ','_') for th in tables[4].find_all('th')]
main_header.append('table')
main_header

notable_fires = []
for i, table in enumerate(tables):
    table_df = {}
    headers = {th.text.lower().strip().replace(' ','_'):index for index,th in enumerate(tables[4].find_all('th'))}

    for row in table.find_all('tr')[1:]:
        notable_f = {}

        for j,feature in enumerate(main_header):
            try:
                notable_f[feature] = row.find_all('td')[headers[feature]].text.strip()
            except:
                notable_f[feature] = np.nan
                
        notable_f['table'] = i
        notable_fires.append(notable_f)
pd.DataFrame(notable_fires).to_csv('./Data/notable_fires_data.csv',index=False)

df = pd.read_csv('./Data/notable_fires_data.csv')

notable_fires = df.loc[df['table'] == 4]

notable_fires.head()

for col in notable_fires.columns:
    print(col)
    notable_fires[col] = notable_fires[col].map(lambda x: re.sub(pattern='(\[\d+\])+$',repl='',string=x))

notable_fires['hectares'] = notable_fires['hectares'].str.replace(",","")
notable_fires['start'] = notable_fires['start'].str.replace(",","")
notable_fires['contained'] = notable_fires['contained'].str.replace(",","")
notable_fires['acres'] = notable_fires['acres'].str.replace(",","")
notable_fires['county'] = notable_fires['county'].str.replace(",","")
notable_fires = notable_fires.replace(np.nan, '', regex=True)
notable_fires = notable_fires.drop(['table'], axis = 1)
notable_fires = notable_fires.drop(['ref'], axis = 1)

notable_fires.head()

notable_fires.to_csv('../data/notable_fires_data.csv',index=False)

### SUMMARY

- This notebook scrapes notable California fire information from a wikipedia page. The columns include the name of the fire, county, acres, hectares, start, contained, and any notes relative to the fire.
