==> 05_Evacuation_Tweet/code/01_Historic_and_Live_Twitter_Scraper.ipynb <==
# Historic and Live Twitter Scraper

In this notebook, we will use the [Notable Fires data](../data/notable_fires_data.csv) from Wikipedia pulled in notebook [00_Wiki_Scrapes_Notable_Fires](./00_Wiki_Scrapes_Notable_Fires.ipynb). We will use the county and date range from Notable Fires in the state of California to decide which California Transit Twitter accounts to pull from and during which date ranges.

This notebook also makes use of the Twitter API: [**Tweepy**](https://www.tweepy.org/) and another Twitter wrapper called [GetOldTweets3 by Jefferson-Henrique](https://github.com/Jefferson-Henrique/GetOldTweets-python/tree/master/got3).

Use of the Tweepy API requires the user to have a twitter handle and a list with specified users to run. In this case, our twitter handle is [@giffordtompkins](https://twitter.com/giffordtompkins) and our list was called [cal-trans-official](https://twitter.com/giffordtompkins/lists/cal-trans-official). Users may specify multiple lists if they wish in the `"TWITTER_LISTS"` argument of the [credential file](../creds/samples/twitter-credentials-sample.json).

## OBJECTIVE 
The goal of this notebook is to pull a historical body of data in the form of tweets from California Transit twitter accounts to build our model upon.

import tweepy
import jsonpickle
import json
import datetime
import GetOldTweets3 as got
import time
import regex as re
import pandas as pd

## Load Tweepy Credentials
You can acquire a Twitter API Credential to be used with Tweepy [here](https://developer.twitter.com/en/docs/basics/getting-started).

*-- The following code is adapted from [Temple Moore](https://github.com/templecm4y/GA-Twitter-Road-Closures-Client-Project/blob/master/code/1_Tweepy-Historical-Tweets-Temple.ipynb).*

# load Twitter API credentials
with open('../creds/samples/twitter-credentials.json') as cred_data:
    info = json.load(cred_data)
    handle = info['TWITTER_HANDLE']
    consumer_key = info['CONSUMER_KEY']
    consumer_secret = info['CONSUMER_SECRET']
    access_token = info['ACCESS_TOKEN']
    access_secret = info['ACCESS_SECRET']
    twitter_lists = info['TWITTER_LISTS']

# Authenticate Tweepy connection with Twitter
auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_secret)


api = tweepy.API(auth)
user = api.get_user(handle)

try:
    api.verify_credentials()
    print("Authentication OK")
except:
    print("Error during authentication")

# Confirm our connection and our access bandwidth.
user = api.me()
username = user.screen_name
print (user.name)
print(username)
print(api.rate_limit_status()['resources']['search'])

# Instantiate list of twitter accounts to scrape for historical tweets.
my_accounts = []

# Pull account names from twitter lists.
for twitter_list in twitter_lists:
    for member in tweepy.Cursor(api.list_members, username, twitter_list).items():
        my_accounts.append(member.screen_name)

# Show list
my_accounts

# Pull CalTrans accounts specifically
cal_trans_accts = []
for account in my_accounts:
    if re.findall('\d+',account) != []:
        cal_trans_accts.append((int(re.findall('\d+',account)[0]),account))
cal_trans_accts
cal_trans_df = pd.DataFrame(cal_trans_accts,index=range(len(cal_trans_accts)),columns=['district','account'])
cal_trans_df.sort_values(by='district')

# Establishing GetOldTweets3 Paramaters
Using the [CalTrans districts data](../data/scrapes/cal-trans-districts.csv) and [notable fires](../data/scrapes/notable_fires_data.csv) datasets from notebooks [00 WikiScrapes CalTrans Districts](00_Wiki_Scrapes_Caltrans_Districts.ipynb) and [00 WikiScrapes Notable Fires](00_Wiki_Scrapes_Notable_Fires.ipynb), we can establish the twitter search parameters to pull the tweets we need to build our historical data set.

districts = pd.read_csv('../data/scrapes/counties_district_scrape.csv')
districts = districts.iloc[13:25,0:2]
districts.columns = ['district', 'counties']
districts

# Split counties string into list and create 1-to-1 dictionary of counties to CalTrans Districts
districts['counties'] = districts['counties'].map(lambda x: [s.strip() for s in x.split(',')])

districts_list = districts['district']
counties = {}
for i in range(len(districts_list)):
    counties_list = districts['counties'].values[i]
    for county in counties_list:
        counties[county] = int(districts_list.values[i])
    
print(counties)
    

! ls ../data/scrapes

# Import fire data.
fires = pd.read_csv('../data/scrapes/notable_fires_data.csv')
fires.head()

fires['district'] = fires['county'].map(counties)
fires

# Lines with multiple counties need to be fixed, individually.
fires[fires['district'].isnull()]['county']

fires['district'] = fires['district'].fillna({
    32:counties['Amador'],
    33:counties['Napa'],
    43:counties['Ventura'],
    47:counties['Glenn'],
    49:counties['Los Angeles']
})

extra_Lake = dict(fires.iloc[33])
extra_Lake['district'] = counties['Lake']
fires = fires.append(extra_Lake,ignore_index=True)

extra_Mendocino = dict(fires.iloc[47])
extra_Mendocino['district'] = counties['Mendocino']
fires = fires.append(extra_Mendocino,ignore_index=True)

fires;

fires = fires.merge(cal_trans_df, how='inner', on='district')
fires.head()

# Convert datetimes to twitter-friendly formats.
fires['start'] = pd.to_datetime(fires['start']).dt.strftime("%Y-%m-%d")
fires['contained'] = pd.to_datetime(fires['contained']).dt.strftime("%Y-%m-%d")
fires.head()

# Create the a list of dictionaries to pass into the GOT3 query field.
tweet_params = []
for index, row in fires[['account','start','contained']].iterrows():
    tweet_params.append(dict(row))
tweet_params[0:2]

## Create Functions for Pulling Tweets
We will create a few functions here for creating a tweet dictionary item and then pulling the batch of tweets from GetOldTweets3.

def create_tweet(tweet):
    '''
    Extract relevant information from the tweets collected by GetOldTweets3 and converts into a dicitonary.
    
    Parameters
    ----------
    tweet (json):
        The tweet object returned from GetOldTweets3
        
    Returns
    -------
    tweet_dict (dict):
        dictionary object with: 
            `id`, 
            `username`, 
            `date`, 
            `text`, 
            `hashtags`, 
            `geo` (geolocation of tweet, to be used when live civilian tweets are incorporated in future iterations), 
            `type` ("official", to be distinguished from civilian tweets in future iterations)
    '''
    tweet_dict = {}
    tweet_dict['id'] = tweet.id
    tweet_dict['username'] = tweet.username
    tweet_dict['date'] = tweet.date
    tweet_dict['text'] = tweet.text
    tweet_dict['hashtags'] = tweet.hashtags
    tweet_dict['geo'] = tweet.geo
    tweet_dict['type'] = 'official' 
    return tweet_dict

def grab_batch(batch_params, max_tweets=1000):
    '''
    This function grabs a batch of tweets of size `max_tweets` based on `batch_params` and appends them to a passed `append_list`.
    
    Parameters
    ----------
    batch_params (dict):
        dicitonary of query paramaters for the batch pull.
    max_tweets (int): default=1000
        maximum number of tweets to pull in each batch.
        
    Returns
    -------
    append_list (list):
        returns the appended list.
    '''
    # Instantiate list of tweets to return 
    tweet_list = []
    # establish our twitter criteria
    tweetCriteria = got.manager.TweetCriteria().setUsername(batch_params['account']).setSince(batch_params['start']).setUntil(batch_params['contained']).setMaxTweets(max_tweets)   

    # query GetOldTweets3
    tweets = got.manager.TweetManager.getTweets(tweetCriteria)

    # create dictionary from tweet object and append to list
    for tweet in tweets:
        tweet_list.append(create_tweet(tweet))
    time.sleep(3)

    return tweet_list

historical_tweets = []
for tweet_param in tweet_params:
    historical_tweets += grab_batch(tweet_param)

# Write historical tweets dataframe into a csv
pd.DataFrame(historical_tweets).to_csv('../data/tweets/historical_tweets.csv', index=False)

# Grab Live Tweets
This will pull the most recent tweets from all CalTrans twitter accounts.

def grab_current_tweets(twitter_account):
    tweets_list = []
    # establish our twitter criteria
    tweetCriteria = got.manager.TweetCriteria().setUsername(twitter_account).setMaxTweets(200)   

    # query GOT
    tweets = got.manager.TweetManager.getTweets(tweetCriteria)

    # create dictionary from tweet object
    for tweet in tweets:
        tweets_list.append(create_tweet(tweet))
    time.sleep(3)

    # added tweet dictionary to list
    return tweets_list

current_tweets = []
for account in my_accounts:
    current_tweets += grab_current_tweets(account)

pd.DataFrame(current_tweets).to_csv('../data/tweets/current_tweets.csv', index=False)

# Summary
In this notebook we established a historical corpus as well as a up-to-date library of CalTrans Tweets.
