==> 05_Evacuation_Tweet/code/02_Tweet_and_spaCy_Exploration.ipynb <==
# Tweet and spaCy EDA
This notebook looks at the capabilities of spaCy in extracting location in formation from our live tweets.

import jsonpickle
import json
import datetime
import time

import regex as re
import pandas as pd
import spacy

live_tweets = pd.read_csv('../data/tweets/current_tweets.csv')

live_tweets.head()

nlp = spacy.load("en_core_web_sm")

live_tweets[live_tweets['username']=="CaltransDist6"]['text']

# Training spaCy
Training spaCy is an extensive process. It is possible to add known entities to its basic vocabulary. To do this, we scraped every street name in California, but the syntax for appending those terms to the spaCy library was beyond what our time allowed.

In future iterations, we would either need to build up this process to create a robust library of street names, interstates, state routes, exits and cross-streets for whatever region we are looking at, or we would need to find a workaround that did not rely on spaCy.

# spaCy with its basic english package, is able to recognize locations like 
# "California" and "Asia" it was unable to pull streenames for us.
doc = nlp(live_tweets['text'][1])

spacy.displacy.render(doc, style='ent', jupyter=True)

streets = pd.read_csv('../data/scrapes/california_street_names.csv')

## spaCy training example from website.
Example of training spaCy's named entity recognizer, starting off with an
existing model or a blank model.

For more details, see the documentation:
* Training: https://spacy.io/usage/training
* NER: https://spacy.io/usage/linguistic-features#named-entities

Compatible with: spaCy v2.0.0+
Last tested with: v2.1.0
"""

```python
from __future__ import unicode_literals, print_function

import plac
import random
from pathlib import Path
import spacy
from spacy.util import minibatch, compounding

TRAIN_DATA = listo

@plac.annotations(
    model=("Model name. Defaults to blank 'en' model.", "option", "m", str),
    output_dir=("Optional output directory", "option", "o", Path),
    n_iter=("Number of training iterations", "option", "n", int),
)
def main(model=None, output_dir=None, n_iter=100):
    """Load the model, set up the pipeline and train the entity recognizer."""
    if model is not None:
        nlp = spacy.load(model)  # load existing spaCy model
        print("Loaded model '%s'" % model)
    else:
        nlp = spacy.blank("en")  # create blank Language class
        print("Created blank 'en' model")

    # create the built-in pipeline components and add them to the pipeline
    # nlp.create_pipe works for built-ins that are registered with spaCy
    if "ner" not in nlp.pipe_names:
        ner = nlp.create_pipe("ner")
        nlp.add_pipe(ner, last=True)
    # otherwise, get it so we can add labels
    else:
        ner = nlp.get_pipe("ner")

    # add labels
    for _, annotations in TRAIN_DATA:
        for ent in annotations.get("entities"):
            ner.add_label(ent[2])

    # get names of other pipes to disable them during training
    other_pipes = [pipe for pipe in nlp.pipe_names if pipe != "ner"]
    with nlp.disable_pipes(*other_pipes):  # only train NER
        # reset and initialize the weights randomly – but only if we're
        # training a new model
        if model is None:
            nlp.begin_training()
        for itn in range(n_iter):
            random.shuffle(TRAIN_DATA)
            losses = {}
            # batch up the examples using spaCy's minibatch
            batches = minibatch(TRAIN_DATA, size=compounding(4.0, 32.0, 1.001))
            for batch in batches:
                texts, annotations = zip(*batch)
                nlp.update(
                    texts,  # batch of texts
                    annotations,  # batch of annotations
                    drop=0.5,  # dropout - make it harder to memorise data
                    losses=losses,
                )
            print("Losses", losses)

    # test the trained model
    for text, _ in TRAIN_DATA:
        doc = nlp(text)
        print("Entities", [(ent.text, ent.label_) for ent in doc.ents])
        print("Tokens", [(t.text, t.ent_type_, t.ent_iob) for t in doc])

    # save model to output directory
    if output_dir is not None:
        output_dir = Path(output_dir)
        if not output_dir.exists():
            output_dir.mkdir()
        nlp.to_disk(output_dir)
        print("Saved model to", output_dir)

        # test the saved model
        print("Loading from", output_dir)
        nlp2 = spacy.load(output_dir)
        for text, _ in TRAIN_DATA:
            doc = nlp2(text)
            print("Entities", [(ent.text, ent.label_) for ent in doc.ents])
            print("Tokens", [(t.text, t.ent_type_, t.ent_iob) for t in doc])
```

listo = []
for index, item in TRAIN_DATA.iterrows():
    listo.append((item['TEXT'], {"entities": [(item['START'], item['END'], "STREET")]}))
listo

TRAIN_DATA = streets[['TEXT', 'ENTITY', 'START','END', 'LABEL']]

## Summary
spaCy has a very nice display feature where you can see the color-coded entities. Again, it is pulling large districts and counties, but is not finding street names. We would like to continue exploring this method, but may need to move on to new methods.
