==> 05_Evacuation_Tweet/code/03_spaCy.ipynb <==
# spaCy 

spaCy is a library used for Natural Language Processing(NLP) in Python. It uses Named Entity Recognition(NER), word vectors and more in order to recognize locations more accurately by detecting certain key words like "at" or "in" before the location. by modifying the strings, SpaCy becomes more effective and gathers more location information.

We specifically call spacy to highlight Countries, cities, and States (GPE), non-tagged locations, mountain ranges, bodies of water (LOC), and buildings, airports, highways, and bridges (FAC)

This notebook is to experiment with Spacy using a subset of our tweet data to attempt to extract locations (cross streets or coordinates) from tweets. spaCy does not require an API key and has a tremendous active database for self education on spaCy. More information on spaCy can be found at https://spacy.io/api.


import pandas as pd
import numpy as np
import re
import spacy
import string
import datetime

import spacy
from spacy.matcher import Matcher
from spacy.tokens import Span

from spacy import displacy



# Reading in our subset data as df. 

df = pd.read_csv("../data/tweets/historical_tweets.csv")

# Calling our data to ensure we read in correctly.  
df

# calling the data within the text column.
df['text']

# Checking the amount of tweets within our subset to ensure a good amount of data is present.
# Print DF shape
print(df.shape)

# Show head of tweets
df.head()

# Create new column for Spacy tweet text.
df['modified_text'] = ''
df['location'] = ''

# Show modified DF
df.head(2)

# dict from DC past cohert project (defining the abbreviations of roads to aid Spacy in recognizing locations)
format_dict = {"hwy": "highway ",
            "blvd": "boulevard",
            " st": "street",
           "CR ": "County Road ",
           "SR ": "State Road",
           "I-": "Interstate ",
           "EB ": "Eastbound ",
           "WB ": "Westbound ",
           "SB ": "Southbound",
           "NB ": "Northbound",
           " on ": " at ",
           " E ": " East ",
           " W ": " West ",
           " S ": " South",
           " N ": " North",
           "mi ": "mile ",
           "between ": "at ",
           "Between ": "at ",
           " In ": " in",
           " in ": " at "}

def spacy_cleaner(df, col, word_dict):
    modified_text = "At " + df[col].replace(word_dict, regex=True)
    modified_text = modified_text.str.title()
    return modified_text

df['modified_text'] = spacy_cleaner(df, 'text', format_dict)

df['modified_text']

def spacy_loc(df, text, geo_column):  # here, we are applying SpaCy to the --
                                     # text column to attempt to extract any location info. 
    
    for i in range(len(df)):  # search entire length of dataframe
        
        #instantiate spacy model
        nlp = spacy.load("en_core_web_sm") # en_core_web_sm is setting language to English
        
        # create documewnt from modified text column
        doc = nlp(df['text'].iloc[i]) #attaching nlp to df to text column, then --
                                           # calling back to doc
        
        locations = set()

## still needs to be defined

# Code below used from Location-Extraction notebook from DC GA-Twitter-Road-Closures-Client-Project
        # loop through every entity in the doc
        for ent in doc.ents:
            
            # find entities labelled as places
            if (ent.label_=='GPE') or (ent.label_=='FAC') or (ent.label_ == 'LOC'):
                
                # put locations in a set
                locations.add(ent.text)
                df[geo_column].iloc[i] = locations
                
    return df[geo_column]

df.head()

df['location']

spacy_loc(df,'tweet','location')
df['location'].head()

## Summary

We attempted to implement spaCy into our project to extract street names in order to plot specific latitude and longitude points within Google and/or Here.com's API’s. 
We noticed SpaCy required a significant amount of model training in order to recognize the street names let alone extract this information and communicate it effectively to Google or Here.com. With thousands of streets we would need to manually input from our Tweets, we chose to look for alternatives.
